sphinx_agent: Move contents of __init__.py to server.py
This allows the main language server to share the type definitions in
types.py without requiring sphinx to be importable.
