+++ lib/esbonio/esbonio/sphinx_agent/__main__.py
@@ -1,5 +1,5 @@
 import asyncio
 
-from esbonio.sphinx_agent import main
+from esbonio.sphinx_agent.server import main
 
 asyncio.run(main())
